requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7 